+++ app/src/main/java/uk/co/group35/app/user/DBModels/Host.java
@@ -2,12 +2,12 @@ package uk.co.group35.app.user.DBModels;
 
 public class Host extends User {
 
-    public Host(Long id, String username, String password) {
+    public Host(Integer id, String username, String password) {
         super(id, username, password);
     }
 
-    public void setupEvent(int eventID, String event_name, String start_date, String end_date) {}
+    public void setupEvent(Integer eventID, String event_name, String start_date, String end_date) {}
 
-    public void getFeedback(int eventID) {}
+    public void getFeedback(Integer eventID) {}
 
 }
